scripts/instagram_scraper/requirements.txt
instaloader==4.13.1
anthropic>=0.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
